tf/requirements-gpu.txt
antlr4-python3-runtime==4.7
jupyter==1.0.0
numpy==1.16.4
pandas==0.23.4
scikit-learn==0.19.2
scipy==1.1.0
tensorflow-gpu==1.15.2
requests
